refactor: move clone into util folder on root level
