+++ src/core/util/index.ts
@@ -1,5 +1,4 @@
 export * from './add-to-buildable-if-processor';
-export * from './clone';
 export * from './create-buildable';
 export * from './create-builder.fn';
 export * from './create-processor.fn';
