+++ src/util/clone.ts
