+++ src/util/index.ts
@@ -1,3 +1,4 @@
+export * from './clone';
 export * from './is-array';
 export * from './is-date.instance';
 export * from './is-defined';
